organizando o código com mais classes (classe Main, classe floatAritmetica,classe floatAritmeticaTest)
